REF reafctor a bit more
