+++ tests/test_mamba_solvable.py
@@ -13,11 +13,13 @@ from conda_forge_feedstock_check_solvable.mamba_solver import (
     FakeRepoData,
     MambaSolver,
     _mamba_factory,
+    virtual_package_repodata,
+)
+from conda_forge_feedstock_check_solvable.utils import (
     _norm_spec,
     apply_pins,
     is_recipe_solvable,
     suppress_conda_build_logging,
-    virtual_package_repodata,
 )
 
 FEEDSTOCK_DIR = os.path.join(os.path.dirname(__file__), "test_feedstock")
